apps/everest/screenshots_generator/requirements.txt
async-generator==1.10
attrs==23.1.0
boto3==1.26.160
botocore==1.29.160
certifi==2023.5.7
charset-normalizer==3.1.0
exceptiongroup==1.1.1
h11==0.14.0
idna==3.4
jmespath==1.0.1
logging==0.4.9.6
numpy==1.25.0
outcome==1.2.0
packaging==23.1
pandas==2.0.2
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
s3transfer==0.6.1
selenium==4.10.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.3
tzdata==2023.3
urllib3==1.26.16
webdriver-manager==3.8.6
wsproto==1.2.0
